Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -2,6 +2,6 @@
 int main() {
     char ch;
     scanf("%c",&ch);
-    printf("ASCII Value: %d",ch+32);
+    printf("ASCII Value: %d",(int)ch+32);
     return 0;
 }
